Allow toci_instack_multinode.sh to run single node
Rename toci_instack_multinode.sh to toci_instack_osinfra.sh and
edit the script to run multi-node actions only on nodes that contains
something in /etc/nodepool/sub_nodes.

+++ toci_instack_osinfra.sh
@@ -66,41 +66,42 @@ source $TRIPLEO_ROOT/tripleo-ci/deploy.env
 
 deploy_env=$(cat $TRIPLEO_ROOT/tripleo-ci/deploy.env)
 
-for ip in $(cat /etc/nodepool/sub_nodes); do
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        sudo yum -y install wget
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        sudo wget http://$MY_IP:8766/current/delorean-ci.repo -O /etc/yum.repos.d/delorean-ci.repo
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        sudo sed -i -e \"s%baseurl=.*%baseurl=http://$MY_IP:8766/current/%\" /etc/yum.repos.d/delorean-ci.repo
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        sudo sed -i -e 's%priority=.*%priority=1%' /etc/yum.repos.d/delorean-ci.repo
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        sudo mkdir -p $TRIPLEO_ROOT/tripleo-ci
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        /bin/bash -c "echo \"$deploy_env\" > deploy.env"
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        sudo cp deploy.env $TRIPLEO_ROOT/tripleo-ci/deploy.env
-done
+# This will remove any puppet configuration done by infra setup
+sudo yum -y remove puppet facter hiera
 
 # TODO: remove later, this is for live debugging
 sudo cat /etc/nodepool/*
 
-# This will remove any puppet configuration done by infra setup
-sudo yum -y remove puppet facter hiera
+if [ -s /etc/nodepool/sub_nodes ]; then
+    for ip in $(cat /etc/nodepool/sub_nodes); do
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            sudo yum -y install wget
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            sudo wget http://$MY_IP:8766/current/delorean-ci.repo -O /etc/yum.repos.d/delorean-ci.repo
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            sudo sed -i -e \"s%baseurl=.*%baseurl=http://$MY_IP:8766/current/%\" /etc/yum.repos.d/delorean-ci.repo
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            sudo sed -i -e 's%priority=.*%priority=1%' /etc/yum.repos.d/delorean-ci.repo
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            sudo mkdir -p $TRIPLEO_ROOT/tripleo-ci
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            /bin/bash -c "echo \"$deploy_env\" > deploy.env"
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            sudo cp deploy.env $TRIPLEO_ROOT/tripleo-ci/deploy.env
+    done
+    $TRIPLEO_ROOT/tripleo-ci/scripts/tripleo.sh --multinode
 
-$TRIPLEO_ROOT/tripleo-ci/scripts/tripleo.sh --multinode
-
-# This needs to be done after the --multinode setup otherwise /etc/hosts will
-# get overwritten
-hosts='127.0.0.1   localhost localhost.localdomain localhost4 localhost4.localdomain4
-::1         localhost localhost.localdomain localhost6 localhost6.localdomain6'
-for ip in $(cat /etc/nodepool/sub_nodes); do
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        /bin/bash -c "echo \"$hosts\" > hosts"
-    ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
-        sudo cp hosts /etc/hosts
-done
+    # This needs to be done after the --multinode setup otherwise /etc/hosts will
+    # get overwritten
+    hosts='127.0.0.1   localhost localhost.localdomain localhost4 localhost4.localdomain4
+    ::1         localhost localhost.localdomain localhost6 localhost6.localdomain6'
+    for ip in $(cat /etc/nodepool/sub_nodes); do
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            /bin/bash -c "echo \"$hosts\" > hosts"
+        ssh $SSH_OPTIONS -tt -i /etc/nodepool/id_rsa $ip \
+            sudo cp hosts /etc/hosts
+    done
+fi
 
 # Add a simple system utilisation logger process
 sudo dstat -tcmndrylpg --output /var/log/dstat-csv.log >/dev/null &
